RabbitMQ online ingestor and ILL accessGroups config
