+++ resources/config.sample.json
@@ -53,12 +53,12 @@
   },
   "rabbitmq": {
     "enabled": true,
-    "host": "localhost",
+    "host": "betalogs.ill.fr",
     "port": 5672,
     "virtual_host": "/",
     "queue_name": "scicat_data",
-    "exchange_name": "",
-    "routing_key": "",
+    "exchange_name": "scicat",
+    "routing_key": "nomad.scicat.publish",
     "auto_ack": false,
     "individual_message_acknowledge": true,
     "prefetch_count": 1,
